AG-13042 Move the legacy theme variable check into the ValidationModule
The check is a development-time diagnostic, so it belongs with the other
ones rather than in core. Environment keeps only the two call sites and
the flag gating them, and reaches the check through beans.validation? in
the same way as checkRowEvents and validateIcon - so with no
ValidationModule registered nothing is scanned and getComputedStyle is
never called.

The variable map and the scan move to validation/rules/themeValidations,
the once-per-grid flag and the warning to ValidationService.
